Made parsing public.
Now Parse() and ParseAll() are available.

Also id's can be zero.
