+++ README.md
@@ -3,5 +3,13 @@
  `users` is a `go` module to manage user data for users that can have access to a server.
  The data are stored in a file like the password file in `*nix`.
 
-The following data are stored: user name, hashed password, user id, zero or more group id's, name, time of creration
-and last time of modification.
+The following data are stored: 
+
+	- user name, should be a valid e-mail address
+	- hashed password as generated by [bcrypt](https://pkg.go.dev/golang.org/x/crypto/bcrypt)
+	- user id, a non negative int value
+	- zero or more group id's; each a non negative int value
+	- name, the full name of the user
+	- time of creation
+	- last time of modification
+
