+++ user.go
@@ -3,6 +3,7 @@ package users
 import (
 	"fmt"
 	"slices"
+	"strconv"
 	"strings"
 	"time"
 
@@ -47,34 +48,104 @@ func (u User) Name() string {
 	return u.name
 }
 
-// New returns a new User. userName should be a valid e-mail address,
-// otherwise the user name wil become an empty string.
+// New returns a new User. userName should be a valid e-mail address.
 // Only positive and unique group id's will be accepted.
-// The user id will have a value of zero, which is an invalid value.
-// After putting it in a AllUsers struct, it will be set to a valid one.
+// The user id will have a value of zero. After putting it in a AllUsers
+// struct, it will be set to a unique value.
 // User has a deactivated status.
 func New(userName, name string, groupIds []int) (User, error) {
 	u := User{
+		userName:       strings.TrimSpace(userName),
 		name:           name,
 		hashedPassword: "*",
 		created:        time.Now(),
 	}
-	if !isValidEMailAddress(userName) {
+	u.modified = u.created
+	if !isValidEMailAddress(u.userName) {
 		return u, ErrInvalidUserName
 	}
-	u.userName = userName
-	u.modified = u.created
-	err := u.SetGroups(groupIds)
 
-	return u, err
+	return u, u.SetGroups(groupIds)
 }
 
-// SetGroups sets the group id's. Only positive and unique group id's
+// Parse creates single User instance by parsing a string. The string must be formatted
+// accordingly to the one as returned by String().
+func Parse(s string) (User, error) {
+	u := User{}
+	var err error
+
+	fields := strings.Split(s, ";")
+	if l := len(fields); l < 7 {
+		return u, fmt.Errorf("%w, less than 7 fields found: %d", ErrMissingData, l)
+	}
+	fields = fields[:7]
+
+	for i, fld := range fields {
+		fld = strings.TrimSpace(fld)
+
+		switch i {
+		case 0: // userName, must be a valid e-mail address
+			if err = u.SetUserName(fld); err != nil {
+				return u, fmt.Errorf("%w (%s)", err, fld)
+			}
+
+		case 1: // hashed password
+			u.hashedPassword = fld
+
+		case 2: // user id
+			u.userId, err = strconv.Atoi(fld)
+			if err != nil || u.userId < 0 {
+				return u, fmt.Errorf("%w for user %s: %s", ErrInvalidUserId, u.userName, fld)
+			}
+
+		case 3: // group id's
+			ids := []int{}
+
+			if len(fld) > 0 {
+				gIds := strings.Split(fld, ",")
+				for _, gId := range gIds {
+					id, err := strconv.Atoi(strings.TrimSpace(gId))
+					if err != nil {
+						return u, fmt.Errorf("%w for user %s: %w",
+							ErrInvalidGroupId, u.userName, err)
+					}
+					ids = append(ids, id)
+				}
+			}
+
+			if err = u.SetGroups(ids); err != nil {
+				return u, fmt.Errorf("cannot set group id's for user %s: %w",
+					u.userName, err)
+			}
+
+		case 4: // name
+			u.name = fld
+
+		case 5: // creation time
+			u.created, err = time.Parse(time.RFC3339, fld)
+			if err != nil {
+				return u, fmt.Errorf("%w (creation) for user %s: %w",
+					ErrInvalidTime, u.userName, err)
+			}
+
+		case 6: // modification time
+			u.modified, err = time.Parse(time.RFC3339, fld)
+			if err != nil {
+				return u, fmt.Errorf("%w (modification) for user %s: %w",
+					ErrInvalidTime, u.userName, err)
+			}
+		}
+	}
+
+	return u, nil
+}
+
+// SetGroups sets the group id's. Only non negative and unique group id's
 // will be accepted.
 func (u *User) SetGroups(groupIds []int) error {
 	ids := []int{}
 	for _, id := range groupIds {
-		if id <= 0 {
+		if id < 0 {
 			return fmt.Errorf("%w: (%d)", ErrInvalidGroupId, id)
 		}
 		if !slices.Contains(ids, id) {
@@ -92,6 +163,19 @@ func (u *User) SetGroups(groupIds []int) error {
 func (u *User) SetName(name string) {
 	u.name = name
 	u.modified = time.Now()
+}
+
+// SetPassword stores a hash of the plain password. If succesfull, it
+// returns nil.
+func (u *User) SetPassword(plainPassword string) error {
+	b, err := bcrypt.GenerateFromPassword([]byte(plainPassword), 12)
+	if err != nil {
+		return err
+	}
+
+	u.hashedPassword = string(b)
+	u.modified = time.Now()
+	return nil
 }
 
 // SetUserName sets the user name. If the user name is not a valid e-mail
@@ -116,19 +200,6 @@ func (u *User) SetUserName(uName string) error {
 		u.userName = uName
 	}
 
-	u.modified = time.Now()
-	return nil
-}
-
-// SetPassword stores a hash of the plain password. If succesfull, it
-// return nil.
-func (u *User) SetPassword(plainPassword string) error {
-	b, err := bcrypt.GenerateFromPassword([]byte(plainPassword), 12)
-	if err != nil {
-		return err
-	}
-
-	u.hashedPassword = string(b)
 	u.modified = time.Now()
 	return nil
 }
